requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
asgiref==3.6.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.2.0
backports.zoneinfo==0.2.1
bleach==6.0.0
blinker==1.5
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
cryptography==39.0.2
DateTime==5.1
defusedxml==0.7.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.18
django-allauth==0.52.0
django-crispy-forms==1.14.0
django-extensions==3.2.1
django-summernote==0.8.20.0
dnspython==2.3.0
eventlet==0.33.3
falcon==3.1.1
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
httplib2==0.22.0
icalendar==5.0.5
icalevents==0.1.25
idna==3.4
MarkupSafe==2.1.2
mongoengine==0.27.0
multidict==6.0.4
numpy==1.24.2
oauthlib==3.2.2
pandas==2.0.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
pystache==0.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.2
requests-oauthlib==1.3.1
Rx==3.2.0
six==1.16.0
sqlparse==0.4.3
static==1.1.1
twitchAPI==3.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
webencodings==0.5.1
Werkzeug==2.2.3
yarl==1.8.2
zope.interface==6.0
